agents/consensus_agents.py
"""
Consensus-based categorization system implementing multiple specialized agents
that collaborate to provide accurate and validated repository categorization.
"""
from typing import List, Dict, Any
from models.curation_models import CategoryHierarchy, EnhancedCurationDetails
from models.repo_models import RepoDetails
from agents.base_agent import BaseAgent

class CategoryAgent(BaseAgent):
    """Agent responsible for initial category generation"""
    
    def __init__(self):
        super().__init__("category_agent")
    
    def execute(self, repo: RepoDetails) -> List[CategoryHierarchy]:
        """Generate initial categories for a repository"""
        return self.categorize(repo)
    
    def categorize(self, repo: RepoDetails) -> List[CategoryHierarchy]:
        """Generate initial categories for a repository"""
        # Example implementation - to be enhanced with actual ML/LLM logic
        categories = [
            CategoryHierarchy(
                main_category="AI/ML",  # Initial categorization
                subcategory="Deep Learning",
                confidence=0.9,
                reasoning="Repository contains deep learning frameworks and neural network implementations"
            )
        ]
        return categories

class ReviewAgent(BaseAgent):
    """Agent responsible for reviewing and refining categories"""
    
    def __init__(self):
        super().__init__("review_agent")
    
    def execute(self, repo: RepoDetails, initial_categories: List[CategoryHierarchy]) -> List[CategoryHierarchy]:
        """Review and refine the initial categories"""
        return self.review(repo, initial_categories)
    
    def review(self, repo: RepoDetails, initial_categories: List[CategoryHierarchy]) -> List[CategoryHierarchy]:
        """Review and refine the initial categories"""
        # Example implementation - to be enhanced with actual review logic
        reviewed_categories = []
        for category in initial_categories:
            # Validate and potentially adjust the category
            reviewed_categories.append(CategoryHierarchy(
                main_category=category.main_category,
                subcategory=category.subcategory,
                confidence=min(category.confidence + 0.05, 1.0),  # Slightly adjust confidence
                reasoning=f"Validated: {category.reasoning}"
            ))
        return reviewed_categories

class ValidationAgent(BaseAgent):
    """Agent responsible for validating categories against historical decisions"""
    
    def __init__(self):
        super().__init__("validation_agent")
    
    def execute(
        self, 
        repo: RepoDetails,
        categories: List[CategoryHierarchy],
        historical_data: Dict[str, Any]
    ) -> List[CategoryHierarchy]:
        """Validate categories against historical decisions"""
        return self.validate(repo, categories, historical_data)
    
    def validate(
        self, 
        repo: RepoDetails,
        categories: List[CategoryHierarchy],
        historical_data: Dict[str, Any]
    ) -> List[CategoryHierarchy]:
        """Validate categories against historical decisions"""
        # Example implementation - to be enhanced with historical validation logic
        validated_categories = []
        for category in categories:
            # Compare with historical decisions and validate
            validated_categories.append(CategoryHierarchy(
                main_category=category.main_category,
                subcategory=category.subcategory,
                confidence=category.confidence,
                reasoning=f"Historically consistent: {category.reasoning}"
            ))
        return validated_categories

class SynthesisAgent(BaseAgent):
    """Agent responsible for synthesizing final categorization results"""
    
    def __init__(self):
        super().__init__("synthesis_agent")
    
    def execute(
        self,
        repo: RepoDetails,
        categories: List[CategoryHierarchy],
        agent_decisions: Dict[str, Any]
    ) -> EnhancedCurationDetails:
        """Synthesize final categorization results"""
        return self.synthesize(repo, categories, agent_decisions)
    
    def synthesize(
        self,
        repo: RepoDetails,
        categories: List[CategoryHierarchy],
        agent_decisions: Dict[str, Any]
    ) -> EnhancedCurationDetails:
        """Synthesize final categorization results"""
        # Example implementation - to be enhanced with proper synthesis logic
        return EnhancedCurationDetails(
            tags=[cat.main_category.lower() for cat in categories],
            popularity_score=0.85,  # To be calculated based on repo metrics
            trending_score=0.75,    # To be calculated based on repo metrics
            categories=categories,
            consensus_data={
                "agent_decisions": agent_decisions,
                "synthesis_confidence": 0.95
            }
        )

class ConsensusManager:
    """Manages the multi-agent consensus process for repository categorization"""
    
    def __init__(self):
        self.category_agent = CategoryAgent()
        self.review_agent = ReviewAgent()
        self.validation_agent = ValidationAgent()
        self.synthesis_agent = SynthesisAgent()
        self.historical_data = {}  # To be replaced with actual historical data storage
    
    def get_consensus(self, repo: RepoDetails) -> EnhancedCurationDetails:
        """
        Orchestrate the consensus process to categorize a repository
        
        Args:
            repo: Repository details to categorize
            
        Returns:
            EnhancedCurationDetails with consensus-based categorization
        """
        # Step 1: Initial categorization
        initial_categories = self.category_agent.execute(repo)
        
        # Step 2: Review and refinement
        reviewed_categories = self.review_agent.execute(repo, initial_categories)
        
        # Step 3: Historical validation
        validated_categories = self.validation_agent.execute(
            repo,
            reviewed_categories,
            self.historical_data
        )
        
        # Step 4: Final synthesis
        agent_decisions = {
            "category_agent": {
                "categories": [c.model_dump() for c in initial_categories],
                "confidence": sum(c.confidence for c in initial_categories) / len(initial_categories)
            },
            "review_agent": {
                "categories": [c.model_dump() for c in reviewed_categories],
                "confidence": sum(c.confidence for c in reviewed_categories) / len(reviewed_categories)
            },
            "validation_agent": {
                "categories": [c.model_dump() for c in validated_categories],
                "confidence": sum(c.confidence for c in validated_categories) / len(validated_categories)
            }
        }
        
        return self.synthesis_agent.execute(repo, validated_categories, agent_decisions)
